course-management-app/requirements.txt
# course-management-app/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
flask-cors==5.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
PyJWT==2.10.1
python-dotenv==1.1.0
Werkzeug==3.1.3